Rating fixes (#129)
* Stops from overwriting rating if no value is set

When PATCH-ing the photo data there is no rating set
and so the value gets set to the default 0. This isn't
right, so let's let the ORM set the value to 0 when
creating a recipe and otherwise leave it to whaveter
it was before when updating if no rating is provided.

* Checks to make sure we're uploading an image

When editing a recipe that already has a photo the data.photo
field is populated with the URL of the image. If we're actually
uploading an image it'll be a "File" object instead.

* Limits rating value when creating recipes